try this one

+++ metadata/client.py
@@ -28,9 +28,9 @@ class PMClient(object):
         """Create the object of type based on hash."""
         ret = requests.put('{0}/{1}'.format(self.url, cls_type),
                            data=dumps(set_hash), headers=self.headers)
-        if ret.status_code / 100 == 2:
+        if int(ret.status_code / 100) == 2:
             return True
-        elif ret.status_code / 100 == 5:
+        elif int(ret.status_code / 100) == 5:
             raise PMClientError('Internal Server Error ({0}) {1}'.format(
                 ret.status_code, ret.content))
         else:
@@ -48,9 +48,9 @@ class PMClient(object):
                             params=query_hash,
                             data=dumps(set_hash),
                             headers=self.headers)
-        if ret.status_code / 100 == 2:
+        if int(ret.status_code / 100) == 2:
             return True
-        if ret.status_code / 100 == 4:
+        if int(ret.status_code / 100) == 4:
             return False
         elif ret.status_code / 100 == 5:
             raise PMClientError('Internal Server Error ({0}) {1}'.format(
@@ -63,9 +63,9 @@ class PMClient(object):
         """Get the object of type from query_hash."""
         ret = requests.get('{0}/{1}'.format(self.url, cls_type),
                            params=query_hash, allow_redirects=True)
-        if ret.status_code / 100 == 2:
+        if int(ret.status_code / 100) == 2:
             return loads(ret.content)
-        elif ret.status_code / 100 == 4:
+        elif int(ret.status_code / 100) == 4:
             return {}
         elif ret.status_code / 100 == 5:
             raise PMClientError('Internal Server Error ({0}) {1}'.format(
